Initial version of the git support for Copybara
This version is able to checkout Git repositories branches, tags
or any ref.

The library receives a repository storate path that uses to keep
all the existing repositories. This is an optimization so that
we don't have to fetch everything all the times (This can take
a considerable time for medium size repos).

Another design consideration is that the library forbids to checkout
local branches. So a user should use origin/master instead of master.
This is so that we can track the updates properly without running
merge. We might allow this in the future, but better to keep it
simple for now.

In addition, a wrapper util for Command is included. This utility
does some non-trivial handling:

  - Log all the commands

  - Log to the console in case of verbose option (when we add it)

  - Keep the stderr/stdout in case of a abnormal termination exception

Change-Id: If48170326d68a5c729d94915f4e7905c59d86361
GitOrigin-RevId: 895656940b72ffbbd524b2ad91d4b5265b10c55c