Integration of git repositories in Copybara
This changes makes available git repositories as source of truth for
Copybara. We can track a default ref defined in the config file or
pass an specific one in the command line.

The mechanism for defining repositories matches what we do for
transformers. Custom classes in YAML files. This allow us to have
custom fields and better validation.

As part of this change a mechanism for pushing flags/configuration to
the parsing has been included. This allow us to have flags per
repository. In this case where we store the repositories and the git
executable path.

Change-Id: I1061906875d4c68a6c061c745652fce3b10960c7
GitOrigin-RevId: 1db9d07ad14c48e1b587a395fb1fbca719fab746