Regex transformer is able to do regex transformations
This change introduces a simple version of the transformer
that reads all the file in memory, does the transformation
using java regex and write the result to the file if it is
modified.

Java groups are allowed. In the form of (group) and $N in the
replacement. Where N = the group number.

Change-Id: I2ee7be6a93d218c8a32483cd02af56e993f80e3a
GitOrigin-RevId: 38a37d6458edcc5a8054e87266b67463ca53f0a7