A new transformer that deletes paths recursively
This will allow us to delete third_party directory when
importing.

Now we also initialize the workdir to be used in an init
method called by Main. So that we deal with tmp dir IOException
in one place instead of each folder. All the transformers can
assume that the Path is valid. Some validation has been added
so that if the option is not set (and we rely on the tmp folder
creation) it fails fast.

Future work: I've been thinking lately about bidirectional
transformations. I think we could make Delete folder Bidirectional
for the normal usecase. We would need to keep a file with the deleted
files and the sha1 of the tree or something similar.

Change-Id: If247c149e26cef969e7a0adcde518cf00677acec
GitOrigin-RevId: d4e9d1efdaefb1eb658f46f4f58fced6c8adb7e5