Allow a GitRepository to use any --git-dir
Rather than calculate the --git-dir deterministically when checking
out a reference, we take the gitDir in the constructor. This doesn't
change the behavior yet, but it makes the code a tiny bit cleaner and
this functionality will be required when creating a new commit.

Creating a new commit requires either
  1) creating a tree programatically or
  2) updating the index

To update the index, we need a non-bare repository, which is different
from the "canonical" repository. The canoincal one is bare and stored
in {GitOptions.gitRepoStorage}/{escaped repo URL}.

Change-Id: Idec3f86e2275b5791672f455dacd5b9e444eca82
GitOrigin-RevId: eecef673c9263578c7ac7c754e493a5c97ec8229