Remove flakey tests that assumed GIT_AUTHOR/COMMITTER_IDENT were stable
These values have the date in them and are used to generate the change
ID hash, so the change ID is different depending on when you run it.
A test to verify that the change ID is deterministic would require so
much injection work so as to make the test burdensome.

Change-Id: I3024f3377fc50b5ba7f6326ed680832c36417444
GitOrigin-RevId: e3dc8c68d3c0b86c737f44d62e5cf8c33a82adf2