Create a Bazel macro that creates a test per file
This is the third time I miss a test. It is too easy to forget
to add a test to the BUILD file. So this should run all **/*Test.java
classes minus exclusions.

The drawback is that all of them share the same dependencies, but that
is fine for now since they almost use the same ones. This could be
minimized in the future by creating more specific packages. Also
note that this is not different from deps perspective to have a
single java_test that globs **java files. On the other hand we have
the benefit of quickly seeing what failed.

This is going to be fun for the import. But it is not bad to test
importing Skylark macros.

Change-Id: I3434514b0d8499995d5780692e3c029dd447c577
GitOrigin-RevId: e3b7cb8de1813448a38a96099ad09543563f69c0