Implement Replace transform
This is similar to ReplaceRegex, but automatically reversible. Because
it is automatically reversible, it only needs to be declared once in
the config YAML, and it can be automatically validated as
bidirectional while transforming the code (simply by applying the
transformation twice in both directions and checking the result
matches the original string).

The unit tests are based heavily on the ReplaceRegex unit tests, since
all tests applicable to the latter are applicable to the former.

This will eventually replace the ReplaceRegex transform.

Change-Id: Ifdc0fa6b9dd53eb08b1a4e00d6235432e3cae4e2
GitOrigin-RevId: 2e3ee49b38475330887fc4cc5016e85ee0b40ab0