GitRepository: forward entire environment to subprocess
If we don't do this, then a user with git not installed on their
system's default $PATH is unable to use copybara. We talked through a
variety of approaches and decided that this was the least hassle. Since
this is a tool users are invoking by hand, if they're owned by running
the tool due to environment variables, they're already owned anyway.

Change-Id: Ida29638e52dd20be2df0988c0b841876c12d0b56
GitOrigin-RevId: ee61bb747672d0798fd5e474ff8d5306ff210cf9