Redirect the log to a file
This change modifies the behavior of the logger to log to a file
instead of the console. The main benefit is that we can see what is
going on instead of seeing hundreds of lines of debug.

The strategy chosen for the logging configuration is to hardcode a
default configuration that looks for the right place to log depending
on the OS (For the common ones).

This behavior can be overriden by passing to copybara the flag:

  --jvm_flag=-Djava.util.logging.config.file=LOG_CONFIG_LOCATION

This is the standard way of configuring the log in Java.

Change-Id: I23522edba5e78d8f2413b28dd7b89d515b8177bf
GitOrigin-RevId: ef22bc5c535fd30dda163e828d7c8f27384908f9