Generate commit message outside of the Destination
Destinations should not be the ones generating the commit message.
Instead it should be passed when called. This allow us to have
custom commit messages (so that we don't have to implement this
functionality per destination) or more importantly allow to
process each change individually and be able to set a commit
message derived from the source commit message (Probably
transformed).

On the other hand, Destinations are allowed to modify the commit
message or add new fields if needed. Practical cases of this
include adding the Change-Id label in gerrit or renaming an existing
label (For example origin could have GERRIT_CHANGE_ID and we would
transform it to Change-Id).

Change-Id: I65c7354068faeb9d163a614144c331d0c20c93ec
GitOrigin-RevId: 3ca6232712046a8ff61bfc08bbe0b706c5b4f8e6