Add a global values field to the config
YAML supports references between nodes. This change adds a field to
aggregate all the global variables in one place instead of spreading
them all over the config file when they are first used. Note that
this doesn't force to use this config style, it just make it possible.

Added an unit test for YAML parser that loads a basic config file
with references.

Change-Id: Ib792b62199dd71374992bf1300f0a20ae13e7afb
GitOrigin-RevId: d5e51254906dbfad739c09555be0591c284ffa6e