Detect generic lists mismatch during the parsing
This (ugly, I know...) code is able to detect when elements
of list fields don't fulfill the expected generic type of
the setter methods. For example, if we have:

public void setStrings(List<? extends Transformer>)

and we pass something like:
  - !Transform1
  - 42

It will detect that 42 is an invalid element for the list and
it will show the proper error.

The code is ugly due to how unflexible Snakeyaml is. OTOH this
allow us to focus on the real business and not spread List<?>
+ a check of elements all over the code.

Change-Id: I32f39a6c8143c2a063288add9d94671f10111452
GitOrigin-RevId: 9441cefa17201bfe7e149329bb915cb9b40864c8