Allow overriding the committer name and e-mail in GitDestination
A follow-up change will cause an exception to be thrown if these are
not set. These flags are partly for convenience of users that don't
have Git set up properly or have a different user name and e-mail set
up than they want to use for Copybara.

This also makes the error thrown when user name/e-mail are not set
actually testable, because the unit tests need to be able to set
these settings without changing the environment. Java processes cannot
change the environment, and the unit tests don't have $HOME set, so
we cannot configure Git user name and e-mail by normal means from
inside tests.

Change-Id: Ibd6beb03f527f7d9cb4b231cff9d0e6a9f5efd57
GitOrigin-RevId: 3c98f062983a4ccbe4a2ee8388ad9bc922b36574