ANSI output for errors and progress messages
This change introduces an optional ansi output console
that allows to show errors, warnings and progress messages
with colors.

This changes also reuses the same line when printing progress
messages in sequence (Like Bazel). Making the whole output
less verbose for the success and at the same time giving
some sense of the progress of the transformation.

As part of this change I also fixed some Exception messages and
how we ouput error to make it more legible and informative
(We had some that were totally uninformative).

I also added some progress messages for origins and destinations.
They do a non-trivial amount of work, for example fetch, and we
want to know what is going on while waiting.

One last note about this. The objective of the progress messages
is not to be totally precise or give all the info but describing
at a high level what we are doing.

Change-Id: Ib1628f9cbeda59a69aebf70d6d2797ff13fcfd15
GitOrigin-RevId: 94f1e24d70ce69aabf03df82568ddccafcd28d50