Fix a fetch bug and add changes to git
Fixes a bug in the fetch code when we try to fetch an sha1. That
doesn't work. You can only fetch refspecs (tags, branches, refs
in general are valid refspecs). So for now if a SHA1 is asked we
fetch the default refspec and hope that it is reachable from the
roots. This is not ideal, but fixing it would require to include
a new configuration field "refspec". So I want to avoid that for
now. We might need this for repos that use non-standard refs,
like Gerrit or GH PR.

In addition to the one before I add support for the changes operation
to GitOrigin. This (plus the bug fix) is the previous work
to have iterative workflows. This operation assumes a linear history.
In case of multiple parents it will chose the first one.

This change also introduces Joda Time. Used for doing some date parsing
in a safe way. Once we can use JDK8 API we will remove it and use
standard Java.

Change-Id: I5b8fa8300f814e39666695568ce89759781ac3a5
GitOrigin-RevId: 5c67861d8f2124c449b96402465955d64e0f98f2