Transformation: allow EnvironmentException to be thrown from withOptions
Some transformations want to run a tool or call an RPC
service. Because of that, let's poke a hole here so that failures
trying to contact an RPC server or locate a tool's executable can be
thrown without forcing the use of an unchecked exception or silent
failure.

We do this here rather than on the transform() method because in the
replicate-each-commit mode, we'll want to check for a tool or create
an RPC connection once and re-use it across all commits.

Change-Id: I878b7756f9c66411135886e4926210e457e26645
GitOrigin-RevId: efc2d356f0511f7cffee3b437fd5fb1f78b542d8