Replace Delete transform by a workflow property
This change removes the Delete transform and replaces it with a field
in the workflow. This allow us to have bidirectional transformations
(as far as the files are also excluded for deletion in the inverse
workflow).

As part of this change the usage of workdir for creation of PathMather
has been removed (Except for Replace transform, that I'll do in the
next change). Now we validate the paths but defer the creation until
we have a workdir. So now we keep it as a Function<Path, PathMatcher>
instead of PathMatcher.

This change also discovered a bug in the deletion code that would make
it possible to delete files outside of the workdir (../../file). This
was not detected because the test was wrong.

Added toString to several PathMatchers.

Migrated the tests from DeleteTest to workflow.

Change-Id: I26cac57cfb15c48b558b5e4a8e66c36049c1d106
GitOrigin-RevId: da691693536ec1ae511f5c6a08911b04e4ea06a6