Move excludedDestinationPaths to an option on Workflow
This makes it possible to implement for all destinations. It may be
possible to allow a destination to explicitly not support excluded
paths, e.g.

if (transformResult.hasExcludedDestPaths().isNotEmpty()) throw new UOE();

But we will want to support this for all our destinations currently,
so I haven't added a way to check for this. Instead, I've added a TODO
to GitDestination to support this.

Change-Id: I8a28232ef3b9fc51fa0d70c6a4495c17a54762cd
GitOrigin-RevId: 6ef8a57688f82fc9af1220cdc94d98e6caca8b8a