Authoring support with three different modes.
DEFAULT: Uses the default author (team) for every change in
the destination.
PASS_THRU: Always uses the origin author in the destination.
WHITELIST: Uses the whitelist to map origin authors to
destination ones. If an author is not present in the whitelist,
then the default author is used.

Regarding configuration validation:
- The default author is always required, because is needed for
every mode (default always uses it, pass_thru uses it for
squash workflows, and whitelist for squash workflows and
non-whitelisted users).
- The whitelist mode requires the map to be not empty.
- The other modes don't require the map, but won't fail if it's
present. In this way a user could change the mode without
having to remove the whitelisted authors.

In the next CLs:
- Repace Map<String, String> by Map<String, Author>.
- Have the origin resolve the autor to the destination one.

Change-Id: I7c386b08fe12c16de4e56b631c28b90633bc8a5e
GitOrigin-RevId: 783cb1a4f0ed48454c5df76e45a4a9c722210484