Adds Author class and GitAuthorParser.
Author is a generic <name, email> class with very lenient validation
that can be parsed from the Yaml config. The only validation is on
the email address and only when read from Yaml. The reason is that
we want to use Author for Git authors, too, and "Foo <bar>" is a
valid author in Git.

GitAuthorParser parses/serializes the Git author string from/to
an Author, following the same validation strictness as Git. It will
be used by the GitOrigin, to read the author from a change, and the
GitDestination, to serialize an author into a new change.

AbstractDestination won't have an author field once we start using
authoring in the Workflow config.

Change-Id: I4e25b38dbad1b00ced43f0556e483273ed8856a1
GitOrigin-RevId: 2b950ae698a3d336af0a5a470f96ac28426dd2a9