Search more than one past commit for last RevId in GitDestination
When looking for most recent RevId, we used to only check the commit
at HEAD. In order to support the use case where some non-migration
commits happen (i.e. to files that are excluded for migration), this
change causes the search to proceed from HEAD until a RevId is found.

Change-Id: Ib12e6911fa7dd60ed8cf8d938e5e57870305f460
GitOrigin-RevId: f24fd19a3cc1d94c62da62eb7c3ea4a112b40c35