Allow Destinations to maintain state between writing of revisions
This slightly refactors the Destination interface to allow the
destination to maintain state after a migration of multiple revisions
has begun. This does not actually take advantage of the more flexible
API, although it is potentially helpful with Git - it will be possible
to avoid cloning the baseline from the remote repository multiple
times. We can just re-use the same clone over and over. Doing this
will enable pushing multiple inter-dependent changes to Gerrit that
are pending review. This fix will come in a follow-up change.

Change-Id: I841ef06c076cef22903938e74d748b296e8fd580
GitOrigin-RevId: 3011daf60a879d0e1c2dc7313940c076a66fa44c