Add field for checking if workflows are reversible
This change introduces a field to make the workflow check at the end of
a migration if it is reverisble back to the origin workdir content.

By default it is set to false, except for CHANGE_REQUEST workflows.
Since it is critical for this kind of workflow to be reversible.
The others (ITERATIVE/SQUASH) are not necessarily reveversible, since
a scrubbing step could happen.

Note that this flag can always be set by the users. But I want to
avoid users needing to add more stuff by default to their config
files.

Change-Id: I18ee693aac0298fd13dc0a179b2b3834fa561436
GitOrigin-RevId: 8a1aeef9c850d9c215bb97e7bddb6245777a2a84