Initial version of the Skylark config files
This change introduces the first version of the code that uses Skylark
as an alternative to YAML.

As a side effect of this change I've removed the need for exposing
Sequence for now. Now that we has Skylark we can use "Python" like
lists, and the only remaining use would be reverse. But we can make
it work for lists and single values. So less concepts that a user
has to learn.

Still pending:

  - Treat reverse

  - Bind options to SkylarkModules as we do for YAML so that we can
    generate good documentation.

Change-Id: I29a25abe2e32e95d7e33bad6731fdfce7ebff9d4
GitOrigin-RevId: a5a4a09537d320d0ce8fa2958b8317d6f0c998aa