Authoring using Skylark
This change adds Skylark support for authoring. Instead of exposing
the enums we provide specialized methods for constructing the objects.
This allows us to use different arguments and avoids common mistakes
like passing a non-empty whitelist to another kind of authoring.

The class uses its own namepace since "authoring.whitelisted" is shorter
than "core.authoring_whitelisted" or similar.

The reason for having two Skylark modules ('authoring' and
'authoring_module') is that the represent two different things.
One is the object that we return and use in workflow, etc. and
the other the factory for construction them. If we merge both modules
we would require:

   - An empty constructor since that is a requirement for modules
     with functions that we want to register in Skylark.

   - The methods are avalable on the returned object. So this would be
     valid code: authoring.whitelisted(...).pass_thru(...). Preventing
     us from making default field, or the whitelist field available
     to our users in the future.

Change-Id: Iaac304e699c63e8944352d3109261639f868576b
GitOrigin-RevId: 731f726583c9ce68663cc9d1801cb3c7b5fe5e86