Introduces new OriginalAuthor.
Represents the author of the change in the origin, as opposed to
Author that is the one in the destination.

GitOriginAuthor resolves automatically to an author because it
has all the information needed. But other implementations might
need to resolve (through a mapping file, an external service, etc).

The resolution is only invoked when is strictly necessary:
- Not for "USE_DEFAULT" authoring mode.
- Not for squash workflows (uses the default author)
- Not for non-whitelisted users in whitelist mode

Change-Id: Ie8ffb0749796066e36404941d2933ce06dc5d6ca
GitOrigin-RevId: c73a8af8afcc71e6c31d46aa54793f3aecd38c15