Introduce a validateConfig() method, to perform global validations.
It's invoked always before running the workflow, and also from the
existing validate() method. Also, the workflow will only be run if
the flag --validate is not set.

The current implementation is empty, but we might move validations
from SkylarkParser in the future. This method will be overridden
by the internal Copybara.

Change-Id: Ib073be84cb181049694c3daf889d141017e63468
GitOrigin-RevId: 18a4761f4ba513b56810674927ab1eaa62be907c