Check non-used groups in 'after' only in reverse
This change moves the check of non-used groups in after field to
reverse instead of doing it eagerly. Since workflows that need reverse
and functions, eagerly ask for the reverse, the net effect is the
same when reverse is required in a config file. No matter that
is not in the same workflow.

But this allow us to scrub content from files when reverse is
not used.

This required more changes than I expected. But OTOH I think we want
to move "officially" reverse() function call to the scope of the
config parsing (IOW, never called during run). So it is OK to throw
EvalException.

Change-Id: Idc0ec710737c077084130b416aa07fb37a4fb8de
GitOrigin-RevId: f8f16d9721e77d5b87c2ce38ebd6daf047f7ad6b