Add core.transform function which allows specifying reversal
This allows you to explicitly specify how to run a transformation in
reverse, which supports some corner-case use cases. In particular, a
user is migrating a large script which uses unusual reversal
semantics - the canonical transform sequence is run in the same order
each time, rather than in opposite order for reversal).

Change-Id: Ic019518edbe74b7daad7025e2c764b88fe7a98f4
GitOrigin-RevId: 1366224a20095eb99d7eadcd0ef0b0e11d66c847