Remove barely-used EnvironmentException in favor of IOException
Stop using "EnvironmentException" and use IOException instead. While
IOException is vague and a catch-all for many kinds of checked
exceptions, EnvironmentException was only being thrown in two contexts
(not thrown at all in internal repo) and we were ultimately handling
it the same way as an IOException.

Also, this exception was preventing me from adding Destination.java to
the copybara:base built target since it caused the util package to
depend on :base and resulted in a dependency loop.

This change adds Destination and TransformResult to the copybara:base
target, since the removal of EnvironmentException makes that possible.

Change-Id: I12020091194d8954c78540ddb728101ba4f9f511
GitOrigin-RevId: cc94cd744d3fc84252a2e69c670ff35dc11c3bbf