Inject environment in GeneralOptions.
The main use case for this is to be able to override it in
environments where Git is not installed by default. For now this is
needed for the internal service, as we need to untar Git ourselves
and add it to the path.

This change removes EnvironmentAware interface, as the environment
can be retrieved now from the GeneralOptions. Also the cwd/homeDir
can be overriden with the environment. If they are not present,
we'll fallback to the standard system property.

Change-Id: Ifdaf65c180f200d1de8df53f5b709a855f32365c
GitOrigin-RevId: 88b5801bc48175196fabda36acfc43565422dc69